# static/swagger6.json
  {
  "openapi": "3.0.0",
  "info": {
    "title": "API BCNC Usuarios y Viviendas",
    "description": "API para la gestión de usuarios y viviendas en BCNC",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
 
 "paths": {
    "/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión y obtener un token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "usuario1"
                  },
                  "password": {
                    "type": "string",
                    "example": "contraseña123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token JWT generado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Credenciales inválidas"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Parámetros de solicitud inválidos"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "El campo 'username' es requerido"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Crear un nuevo usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string",
                    "example": "usuario1"
                  },
                  "password": {
                    "type": "string",
                    "example": "contraseña123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado correctamente"
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Parámetros de solicitud inválidos"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "El campo 'password' debe tener al menos 6 caracteres"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Servicio"],
        "summary": "Obtener todos los usuarios",
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{usuario_id}": {
      "get": {
        "tags": ["Servicio"],
        "summary": "Obtener un único usuario por ID",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuario no encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Servicio"],
        "summary": "Actualizar parcialmente un usuario por ID",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado correctamente"
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Parámetros de solicitud inválidos"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "El campo 'email' no es válido"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Servicio"],
        "summary": "Actualizar completamente un usuario por ID",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado completamente"
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Parámetros de solicitud inválidos"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "El campo 'nombre' es requerido"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Servicio"],
        "summary": "Borrar un usuario por ID",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario borrado correctamente"
          },
          "400": {
            "description": "El usuario tiene viviendas asociadas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "El usuario tiene viviendas asociadas"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{usuario_id}/viviendas": {
      "get": {
        "tags": ["Servicio"],
        "summary": "Obtener todas las viviendas de un usuario",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de viviendas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Servicio"],
        "summary": "Crear una nueva vivienda para un usuario",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vivienda creada correctamente"
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{usuario_id}/viviendas/{vivienda_id}": {
      "patch": {
        "tags": ["Servicio"],
        "summary": "Actualizar parcialmente una vivienda de un usuario",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          },
          {
            "name": "vivienda_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vivienda actualizada correctamente"
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Parámetros de solicitud inválidos"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "El campo 'precio' debe ser un número"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Servicio"],
        "summary": "Actualizar completamente una vivienda de un usuario",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          },
          {
            "name": "vivienda_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vivienda actualizada completamente"
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Parámetros de solicitud inválidos"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "El campo 'número de habitaciones' debe ser un número"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Servicio"],
        "summary": "Borrar una vivienda de un usuario",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          },
          {
            "name": "vivienda_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vivienda borrada correctamente"
          },
          "500": {
            "description": "Error del sistema",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}